pyproject.toml
[tool.poetry]
name = "orkhon"
version = "0.1.0"
description = ""
authors = ["Px8Studio <info@px8.studio>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
toolbox-langchain = "^0.5.2"
langchain-google-vertexai = "^2.1.2"
langgraph = "^0.6.10"
google-adk = "^1.16.0"
uv = "^0.9.3"
toolbox-core = "^0.5.2"
# Kiota dependencies for HTTP clients
microsoft-kiota-abstractions = "^1.7.0"
microsoft-kiota-http = "^1.7.0"
microsoft-kiota-serialization-json = "^1.7.0"
microsoft-kiota-serialization-text = "^1.7.0"
microsoft-kiota-serialization-form = "^1.7.0"
microsoft-kiota-serialization-multipart = "^1.7.0"
httpx = "^0.28.1"
sqlalchemy = "^2.0.44"
psycopg2-binary = "^2.9.11"
pandas = "^2.3.3"
pyarrow = "^21.0.0"
python-dotenv = "^1.0.0"
google-cloud-storage = ">=2.18.0,<3.0.0"
pydantic = "^2.10.6"
pyyaml = "^6.0.1"
google-generativeai = "^0.8.5"
immutabledict = "^4.2.2"
sqlglot = "^27.29.0"
db-dtypes = "^1.4.3"
regex = "^2025.11.3"
tabulate = "^0.9.0"
matplotlib = "^3.10.7"


[tool.poetry.group.dev.dependencies]
ruff = "^0.14.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
